==> src/1_dengue_raw_curated.ipynb <==
from parameters import get_parameters
import pandas as pd
from pathlib import Path
from unidecode import unidecode

params = get_parameters()

start_year = params["start_year"]
end_year = params["end_year"]
source_raw = params["source_raw"]
source_curated = params["source_curated"]

PROJECT_ROOT = Path().resolve().parent
print(f"Project root: {PROJECT_ROOT}")

df_list = []

for year in range(start_year, end_year + 1):
    path_dengue = f"{PROJECT_ROOT}/{source_raw}/Dengue_{year}.xlsx"
    df_dengue = pd.read_excel(path_dengue, engine="openpyxl")
    df_list.append(df_dengue)

for year in range(start_year, end_year + 1):
    path_severe = f"{PROJECT_ROOT}/{source_raw}/Dengue_grave_{year}.xlsx"
    df_severe = pd.read_excel(path_severe, engine="openpyxl")
    df_list.append(df_severe)

df = pd.concat(df_list, ignore_index=True)
print(df.head())

# Filter by confirmed cases
df_cases = df[df["confirmados"] == 1]

# Filter by infections in Colombia
df_cases = df_cases[df_cases["Pais_ocurrencia"] == "COLOMBIA"]

# Filter by department and municipios (first clean the strings)
df_cases["Departamento_ocurrencia"] = df_cases["Departamento_ocurrencia"].apply(lambda x: unidecode(str(x))).str.strip().str.upper()
df_cases["Municipio_ocurrencia"] = df_cases["Municipio_ocurrencia"].apply(lambda x: unidecode(str(x))).str.strip().str.upper()

df_cases = df_cases[df_cases["Departamento_ocurrencia"] == "ANTIOQUIA"]
df_cases = df_cases[df_cases["Municipio_ocurrencia"].isin(["MEDELLIN"])]

print(df_cases.head())

# FEC_CON to datetime type
df_cases["FEC_CON"] = pd.to_datetime(df_cases["FEC_CON"]).dt.date
df_cases["FEC_CON"] = pd.to_datetime(df_cases["FEC_CON"], format="%Y-%m-%d")

# Group dataset by infected date and place
df_cases = df_cases.groupby(["FEC_CON", "Municipio_ocurrencia"]).agg({"COD_EVE" : "size"})

# Index as columns
df_cases = df_cases.reset_index(drop=False)

# Change column names
df_cases = df_cases.rename(columns={"FEC_CON":"DATE", "COD_EVE": "CASES", "Municipio_ocurrencia":"MUNICIPIO"})

# Filter by date range
df_cases = df_cases[(df_cases["DATE"] >= f"{start_year}-01-01") & (df_cases["DATE"] <= f"{end_year}-12-31")]

# Change date to the start of the week
date_range = pd.date_range(start=F"{start_year}-01-01", end=f"{end_year}-12-31", freq="W")
df_cases["DATE"] = df_cases["DATE"].apply(lambda d: date_range[date_range <= d].max())
df_cases = df_cases.dropna(subset=["DATE"])

# Group records by week
df_cases = df_cases.groupby(["DATE", "MUNICIPIO"]).agg({"CASES": "sum"})
df_cases = df_cases.reset_index(drop=False)

# Add missing weeks
municipios = df_cases["MUNICIPIO"].unique()
weeks_x_municipios = pd.MultiIndex.from_product([date_range, municipios], names=["DATE", "MUNICIPIO"]).to_frame(index=False)
df_cases = df_cases.merge(weeks_x_municipios, on=["DATE", "MUNICIPIO"], how="right")
df_cases["CASES"] = df_cases["CASES"].fillna(0)

# Change column CASES to int
df_cases["CASES"] = df_cases["CASES"].astype(int)

# Format column MUNICIPIO
df_cases["MUNICIPIO"] = df_cases["MUNICIPIO"].replace("MEDELLIN", "MEDELLÍN")

# Show dataframe info
print("Total of records:", len(df_cases))
print("============================================")
print("Columns and data types:")
print(df_cases.dtypes)

df_cases.head()

pd.to_pickle(df_cases, f"{PROJECT_ROOT}/{source_curated}/dengue.pkl")
