==> src/2_data_curated_conformed.ipynb <==
from parameters import get_parameters
import os
import re
import numpy as np
import pandas as pd
from pathlib import Path

params = get_parameters()

start_year = params["start_year"]
end_year = params["end_year"]
source_curated = params["source_curated"]
source_conformed = params["source_conformed"]

PROJECT_ROOT = Path().resolve().parent
print(f"Project root: {PROJECT_ROOT}")

df_dengue = pd.read_pickle(f"{PROJECT_ROOT}/{source_curated}/dengue.pkl")

climate_variables = [f.replace(".pkl", "") for f in os.listdir(f"{PROJECT_ROOT}/{source_curated}") if f.endswith('.pkl') and f != 'dengue.pkl']
climate_variables = list(set(climate_variables))
climate_variables

climate_data = {}

for var in climate_variables:
    climate_data[var] = pd.read_pickle(f"{PROJECT_ROOT}/{source_curated}/{var}.pkl")

municipios = df_dengue.MUNICIPIO.unique()

for key, df in climate_data.items():
    municipios = np.concatenate((municipios, df.MUNICIPIO.unique()))
municipios = np.unique(municipios)
print("Total of municipios:", len(municipios))
print("List of municipios:\n", municipios)

df_dengue = (
    df_dengue
    .groupby("DATE", as_index=False)
    .agg({"CASES": "sum"})
)

for variable in climate_variables:
    var = variable.lower()

    df = climate_data[var].copy()

    df = df.drop(columns=[c for c in ["GEOMETRY", "LON_LAT"] if c in df.columns], errors="ignore")

    value_vars = [
        col for col in df.columns
        if col not in ["DATE", "ID", "MUNICIPIO"]
    ]

    df = (
        df.melt(
            id_vars=["DATE", "ID", "MUNICIPIO"],
            value_vars=value_vars
        )
        .assign(variable=lambda d: d["variable"] + "_" + d["ID"].astype(str) + "_" + d["MUNICIPIO"])
        .pivot(index="DATE", columns="variable", values="value")
        .reset_index()
    )
    df.columns.name = None
    climate_data[var] = df


for variable in climate_variables:
    var = variable.lower()

    df = climate_data[var]

    # --- TEMPERATURE ---
    if "temperature" in var:
        df = df.apply(lambda x: x - 273.15 if x.name != "DATE" else x)
        print(f"Changed {variable} from K → °to")

    # --- RUNOFFS & PRECIPITATION ---
    elif any(k in var for k in ["runoff", "precipitation"]):
        df = df.apply(lambda x: x * 1000 if x.name != "DATE" else x)
        print(f"Changed {variable} from m → mto")

    # --- WINDS ---
    elif "wind" in var:
        df = df.apply(lambda x: x * 3.6 if x.name != "DATE" else x)
        print(f"Changed {variable} from m/s to km/h")

    else:
        print(f"⚠️ {variable}: no conversion rule found")

    climate_data[var] = df

# Merge all datasets using DATE field
df_total = df_dengue

for key, df in climate_data.items():
    df_total = df_total.merge(df, on=['DATE'], how='left')

pattern = r'^(.*?)_\d+_'
climate_variables_new = sorted(set(re.match(pattern, col).group(1) for col in df_total.columns if re.match(pattern, col)))
climate_variables_new

# Check if there are any missing values
df_total[df_total.isnull().any(axis=1)]

# Seleccionar valor promedio por variable climática
for var in climate_variables_new:
    cols = [c for c in df_total.columns if c.startswith(var)]   
    col_name = f"{var.split('_')[0]}_AVG_MAX"
    df_total[col_name] = df_total[cols].mean(axis=1)
    df_total.drop(columns=cols, inplace=True)

df_total = df_total.set_index('DATE')
df_total = df_total.asfreq('W')

df_total = df_total.sort_index()
df_total = df_total.iloc[:-1]

# Show dataframe info
print("Total of records:", len(df_total))
print("============================================")
print("Columns and data types:")
print(df_total.dtypes)

df_total.head()

pd.to_pickle(df_total, f"{PROJECT_ROOT}/{source_conformed}/data.pkl")
